### 2021-03-24 #36 @skasberger
pyDataverse: a Python module for Dataverse
Submitting Author: Stefan Kasberger (@skasberger)
Package Name: pyDataverse
One-Line Description of Package: An Open Source Python module to work with Dataverse
Repository Link (if existing):   https://github.com/gdcc/pyDataverse

---

pyDataverse is a Python module for Dataverse you can use for:

* accessing the Dataverse API’s
* manipulating and using the Dataverse (meta)data - Dataverses, Datasets, Datafiles

No matter, if you want to import huge masses of data into Dataverse, test your Dataverse instance after deployment or want to make basic API calls: pyDataverse helps you with Dataverse!

pyDataverse is fully Open Source and can be used by everybody.

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences).  Please note any areas you are unsure of: pyDataverse connects with the Dataverse research data repository software. So it accesses scientific knowledge, but also processes and deposits it.

-   Who is the target audience and what are scientific applications of this package?  The tool itself is not for scientific purpose itself, but it supports science. Students use it to get data from the repository to analyse it later on, but the data and metadata itself could also be used for some meta studies.

-   Are there other Python packages that accomplish the same thing? If so, how does yours differ? No.
