### 2021-03-25 #36 @lwasser
thank you for this presubmission @skasberger ! We are just regrouping after a difficult week here in Boulder and the GOOD news  about getting funding. In the meantime, did you have specific  questions related to our process or is this just a check to ensure the package is in scope for us?
