### 2021-04-06 c280249
Fix issue: topicClassVocabURI missing (#126)
The topicClassVocabURI was not used by models.py, which
lead to not importing or exporting the variable.

Resolves: #126
