### 2021-04-06 555a727
Update DOI to release 0.3.1
