+++ src/pyDataverse/docs/source/index.rst
@@ -32,8 +32,8 @@ Release v\ |version|.
 .. image:: https://img.shields.io/badge/code%20style-black-000000.svg
     :target: https://github.com/psf/black
 
-.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.4470150.svg
-   :target: https://doi.org/10.5281/zenodo.4470150
+.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.4664557.svg
+   :target: https://doi.org/10.5281/zenodo.4664557
 
 -------------------
 
