+++ .travis.yml
@@ -16,8 +16,6 @@ matrix:
       env:       TOXENV=coverage
     - python:    3.6
       env:       TOXENV=codecov
-    - python:    3.6
-      env:       TOXENV=coveralls
 
 branches:
   only:
@@ -32,7 +30,6 @@ install:
   - pip install tox-travis
   - pip install coverage
   - pip install codecov
-  - pip install coveralls
   - virtualenv --version
   - easy_install --version
   - pip --version
@@ -42,9 +39,7 @@ script:
   - tox
 
 after_success:
-  - coveralls
   - bash <(curl -s https://codecov.io/bash)
-  - codecov
 
 notifications:
   email:
