+++ setup.cfg
@@ -4,13 +4,6 @@ license_file = LICENSE.txt
 [tool:pytest]
 testpaths = tests
 
-[coverage:run]
-source = wheel
-omit = .tox/*
-
-[coverage:report]
-show_missing = true
-
 [radon]
 cc_min = B
 
