+++ .travis.yml
@@ -40,6 +40,7 @@ script:
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
+  - codecov
 
 notifications:
   email:
