+++ tox.ini
@@ -37,20 +37,6 @@ deps =
 commands =
     pytest tests/ --cov=pyDataverse --cov-report=term-missing --cov-report=xml --cov-report=html
 
-[testenv:codecov]
-description = create report for coverage
-deps =
-    -r{toxinidir}/requirements/tests.txt
-commands =
-    codecov
-
-[testenv:coveralls]
-description = create reports for coveralls
-deps =
-    -r{toxinidir}/requirements/tests.txt
-commands =
-    pytest tests/ --doctest-modules -v --cov=pyDataverse
-
 [testenv:pylint]
 description = pylint for linting
 deps =
