+++ .travis.yml
@@ -14,8 +14,6 @@ matrix:
       env:       TOXENV=docs
     - python:    3.6
       env:       TOXENV=coverage
-    - python:    3.6
-      env:       TOXENV=codecov
 
 branches:
   only:
@@ -39,7 +37,6 @@ script:
   - tox
 
 after_success:
-  - bash <(curl -s https://codecov.io/bash)
   - codecov
 
 notifications:
