+++ tox.ini
@@ -37,6 +37,13 @@ deps =
 commands =
     pytest tests/ --cov=pyDataverse --cov-report=term-missing --cov-report=xml --cov-report=html
 
+[testenv:codecov]
+description = create report for codecov
+deps =
+    codecov
+commands =
+    codecov
+
 [testenv:pylint]
 description = pylint for linting
 deps =
