+++ .travis.yml
@@ -14,6 +14,8 @@ matrix:
       env:       TOXENV=docs
     - python:    3.6
       env:       TOXENV=coverage
+    - python:    3.6
+      env:       TOXENV=codecov
 
 branches:
   only:
@@ -26,6 +28,8 @@ before_install:
 install:
   - pip install --upgrade pip
   - pip install tox-travis
+  - pip install coverage
+  - pip install codecov
   - virtualenv --version
   - easy_install --version
   - pip --version
@@ -36,6 +40,7 @@ script:
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
+  - codecov
 
 notifications:
   email:
