+++ .travis.yml
@@ -39,8 +39,8 @@ script:
   - tox
 
 after_success:
-  - bash <(curl -s https://codecov.io/bash)
   - codecov
+  - bash <(curl -s https://codecov.io/bash)
 
 notifications:
   email:
