+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py36,py37,py38,coverage,coveralls,docs,packaging,dist_install
+envlist = py36,py37,py38,coverage,docs,packaging,dist_install
 skip_missing_interpreters = True
 ignore_basepython_conflict = True
 
@@ -40,9 +40,9 @@ commands =
 [testenv:codecov]
 description = create report for codecov
 deps =
-    codecov
+    codecov>=1.4.0
 commands =
-    codecov
+    codecov -e TOXENV
 
 [testenv:pylint]
 description = pylint for linting
