+++ .travis.yml
@@ -27,7 +27,7 @@ install:
   - pip install --upgrade pip
   - pip install tox-travis
   - pip install coverage
-  - pip install --user codecov
+  - pip install codecov
   - virtualenv --version
   - easy_install --version
   - pip --version
