### 2021-04-07 9bd2f6a
add codecov and update tox
