+++ requirements/lint.txt
@@ -10,9 +10,9 @@ flake8-breakpoint==v1.1.0
 flake8-bugbear==20.1.4
 flake8-comprehensions==3.2.3
 flake8-requirements==v1.3.2
-flake8-return==0.3.0
 flake8-pytest-style==v1.2.1
 pylint
 radon
 mypy==0.790
 autopep8
+Pygments
