+++ setup.cfg
@@ -11,10 +11,13 @@ cc_min = B
 max-line-length = 88
 ignore = E129,E203,E501,W503,S101
 exclude =
-    .tox
-    .egg
+    docs
+    schemas
+    templates
     conf.py
     conftest.py
+    __pycache__
+    .mypy_cache
 
 [pylint]
 max-line-length = 88
@@ -25,3 +28,15 @@ max_line_length = 88
 [pylama:pylint]
 max_line_length = 88
 disable = R
+
+[coverage:run]
+source=pyDataverse
+
+[coverage:report]
+show_missing = true
+
+[coverage:html]
+directory = docs/coverage/html
+
+[coverage:xml]
+output = coverage.xml
