+++ codecov.yml
@@ -1,15 +1,0 @@
-codecov:
-  branch: master
-
-flags:
-  library:
-    paths:
-    - src/pyDataverse/
-  docs:
-    paths:
-    - src/pyDataverse/docs/
-    - HISTORY.rst
-    - CONTRIBUTING.rst
-    - "*.md"
-    - "*.rst"
-    - "*.txt"
