+++ tox.ini
@@ -40,9 +40,9 @@ commands =
 [testenv:codecov]
 description = create report for codecov
 deps =
-    codecov>=1.4.0
+    -r{toxinidir}/requirements/tests.txt
 commands =
-    codecov -e TOXENV
+    codecov
 
 [testenv:pylint]
 description = pylint for linting
@@ -56,47 +56,21 @@ commands =
 deps =
     -r{toxinidir}/requirements/lint.txt
 commands =
-    mypy src/pyDataverse/ setup.py tests/
-
-[flake8]
-max-line-length = 80
-ignore = E129
-exclude =
-    src/pyDataverse/docs/source/conf.py
-    .tox
-    .egg
+    mypy src/pyDataverse/ setup.py
 
 [testenv:flake8]
 description = flake8 for style guide and docstring testing
 deps =
     -r{toxinidir}/requirements/lint.txt
 commands =
-    pip uninstall -y flake8-pytest-style
-    pip uninstall -y flake8-rst-docstrings
-    flake8 src/pyDataverse/
+    flake8 src/pyDataverse
 
-[testenv:flake8_docs]
+[testenv:flake8_docstrings]
 description = flake8 for style guide and docstring testing
 deps =
     -r{toxinidir}/requirements/lint.txt
 commands =
-    pip uninstall -y flake8-blind-except
-    pip uninstall -y flake8-builtins
-    pip uninstall -y flake8-bandit
-    pip uninstall -y flake8-breakpoint
-    pip uninstall -y flake8-bugbear
-    pip uninstall -y flake8-comprehensions
-    pip uninstall -y flake8-requirements
-    pip uninstall -y flake8-return
     flake8 --docstring-convention numpy src/pyDataverse/
-    flake8 --docstring-convention numpy tests/
-
-[testenv:flake8_tests]
-description = flake8 for style guide and docstring testing
-deps =
-    -r{toxinidir}/requirements/lint.txt
-commands =
-    flake8 tests/
 
 [testenv:black]
 description = black for auto-formatting
@@ -133,9 +107,6 @@ recreate = True
 deps =
     -r{toxinidir}/requirements/packaging.txt
 commands =
-    # rm -R {toxinidir}/src/pyDataverse.egg-info
-    # rm -R {toxinidir}/dist
-    # rm -R {toxinidir}/build
     python setup.py sdist bdist_wheel
     twine check dist/*
 
@@ -151,7 +122,7 @@ description = install from test pypi
 skip_install = True
 recreate = True
 commands =
-    pip install --index-url https://test.pypi.org/simple/ pyDataverse
+    pip install -i https://test.pypi.org/simple/ pyDataverse --extra-index-url https://pypi.org/simple pyDataverse
     pip freeze
     python -c "import pyDataverse; print(pyDataverse.__name__); print(pyDataverse.__version__)"
 
