+++ src/pyDataverse/models.py
@@ -1,6 +1,4 @@
 """Dataverse data-types data model."""
-from __future__ import absolute_import
-
 import json
 import os
 
