+++ src/pyDataverse/__init__.py
@@ -4,8 +4,6 @@ Copyright 2019 Stefan Kasberger
 
 Licensed under the MIT License.
 """
-from __future__ import absolute_import
-
 from requests.packages import urllib3
 
 urllib3.disable_warnings()  # noqa
