+++ requirements/docs.txt
@@ -1,9 +1,8 @@
-# Requirements for documentation test (tox)
 -r common.txt
-sphinx==1.7.*
-git+https://github.com/python/python-docs-theme.git#egg=python-docs-theme
 git+https://github.com/pypa/pypa-docs-theme.git#egg=pypa-docs-theme
+git+https://github.com/python/python-docs-theme.git#egg=python-docs-theme
 pydocstyle
-restructuredtext_lint
 pygments
+restructuredtext_lint
 rstcheck
+sphinx==1.7.*
