+++ requirements/lint.txt
@@ -1,18 +1,17 @@
-# Requirements for linting tests (tox)
+autopep8
 black==19.10b0
 flake8==3.7.4
-flake8-docstrings==1.5.0
-flake8-rst-docstrings==v0.0.13
-flake8-blind-except==v0.1.1
-flake8-builtins==1.5.3
 flake8-bandit==v2.1.0
+flake8-blind-except==v0.1.1
 flake8-breakpoint==v1.1.0
 flake8-bugbear==20.1.4
+flake8-builtins==1.5.3
 flake8-comprehensions==3.2.3
-flake8-requirements==v1.3.2
+flake8-docstrings==1.5.0
 flake8-pytest-style==v1.2.1
+flake8-requirements==v1.3.2
+flake8-rst-docstrings==v0.0.13
+mypy==0.790
+Pygments
 pylint
 radon
-mypy==0.790
-autopep8
-Pygments
