+++ requirements/common.txt
@@ -1,2 +1,2 @@
-requests>=2.12.0
 jsonschema>=3.2.0
+requests>=2.12.0
