+++ requirements/development.txt
@@ -1,9 +1,8 @@
-# Requirements for development.
 -r common.txt
--r lint.txt
 -r docs.txt
+-r lint.txt
 -r tests.txt
-wheel
 pre-commit
-twine
 tox
+twine
+wheel
