+++ setup.cfg
@@ -4,9 +4,6 @@ license_file = LICENSE.txt
 [tool:pytest]
 testpaths = tests
 
-[radon]
-cc_min = B
-
 [flake8]
 max-line-length = 88
 ignore = E129,E203,E501,W503,S101
@@ -18,16 +15,11 @@ exclude =
     conftest.py
     __pycache__
     .mypy_cache
+    .tox
+    .egg
 
 [pylint]
 max-line-length = 88
-
-[pylama:pycodestyle]
-max_line_length = 88
-
-[pylama:pylint]
-max_line_length = 88
-disable = R
 
 [coverage:run]
 source=pyDataverse
@@ -40,3 +32,6 @@ directory = docs/coverage/html
 
 [coverage:xml]
 output = coverage.xml
+
+[radon]
+cc_min = B
