+++ .pre-commit-config.yaml
@@ -15,8 +15,12 @@ repos:
     -   id: check-yaml
     -   id: detect-private-key
     -   id: end-of-file-fixer
+    -   id: fix-encoding-pragma
+        args: [--remove]
     -   id: pretty-format-json
         args: [--autofix,--no-sort-keys]
+    -   id: requirements-txt-fixer
+    -   id: trailing-whitespace
 -   repo: https://github.com/psf/black
     rev: 19.10b0
     hooks:
@@ -27,3 +31,8 @@ repos:
     hooks:
     -   id: blacken-docs
         additional_dependencies: [black==19.10b0]
+-   repo: https://github.com/asottile/reorder_python_imports
+    rev: v2.4.0
+    hooks:
+    -   id: reorder-python-imports
+        args: ['--application-directories=.:src', --py36-plus]
