### 2021-04-07 6478ede
fix minor docs issues
