### 2021-04-12 6dddd4f
fix issue: flake8 exlude filenames without dir
