+++ setup.cfg
@@ -3,14 +3,15 @@ license_file = LICENSE.txt
 
 [tool:pytest]
 testpaths = tests
+addopts = -v
 
 [flake8]
 max-line-length = 88
 ignore = E129,E203,E501,W503,S101
 exclude =
-    docs
-    schemas
-    templates
+    src/pyDataverse/docs
+    src/pyDataverse/schemas
+    src/pyDataverse/templates
     conf.py
     conftest.py
     __pycache__
