### 2021-04-12 cc06022
small fixes in requirements
