+++ requirements/development.txt
@@ -1,8 +1,7 @@
 -r common.txt
 -r docs.txt
 -r lint.txt
+-r packaging.txt
 -r tests.txt
 pre-commit
-tox
-twine
 wheel
