+++ requirements/common.txt
@@ -1,2 +1,4 @@
-requests>=2.12.0
+aiohttp==3.7.3
+asyncio==3.4.3
 jsonschema>=3.2.0
+requests>=2.12.0
