### 2021-04-30 #36 @NickleDave
Thank you @lwasser -- happy to act as editor

@skasberger `pyDataverse` looks like a perfect fit for pyOpenSci and I can see the package is already well-developed and used. Hope you will go ahead and raise an issue here for a full submission

One point I did want to bring up:
under "other packages" you put "no" but there is of course https://github.com/IQSS/dataverse-client-python (with a link to `pyDataverse` in their README).
Even if that package is no longer actively maintained, we should be sure to acknowledge all their hard work. Especially because the developers and maintainers would be potential reviewers. Please do include links to that package in any submission, along with a comparison of functionality
