### 2022-06-14 e21e3a3
fix newer dataverse compatibility by adding content-type=application/json header
 - able to add content-type to post_request
 - set content-type to application/json for strings

based on https://github.com/JR-1991/pyDataverse/commit/0fcfcd3fbc6bf1aec869899f715a51dca25e91be
