### 2022-06-21 0134201
Made request streamable tobe able to get their raw data
