### 2022-06-27 04f2c8b
Fix edit_dataset_metadata docstring
