### 2022-08-13 7052d0a
Fix docstring of replace_datafile
The identifier needed here is that of the to be replaced file, not an
identifier of the dataset.
